Change project structure
Add bin directory and `make build` in order to build the project's
binary inside bin/

Rename cmd directory to commands
