+++ .gitignore
@@ -1,2 +1,2 @@
 # Binary file
-flotta-dev-cli
+bin
